# openapi.yaml
openapi: 3.1.0
info:
  title: Intella API
  version: 0.7.0
  description: >
    Single-user, self-hosted adaptive coach: training, meal planning, and a
    store-agnostic grocery list. Every generator follows the same
    rules -> LLM -> validator pattern and persists the input constraints that
    produced its output (reproducibility + explainability).

    Core entities (Profile, Goal, TrainingProfile, DietProfile) reflect the
    v0.7 preflight resolutions (see Intella_Preflight_Resolutions.md).

    This spec is the single source of truth; the TypeScript web client and the
    Swift iOS client are both generated from it.
servers:
  - url: http://localhost:8787
    description: Local desktop host
  - url: https://intella.your-tailnet.ts.net
    description: Private access over Tailscale (MagicDNS)
security:
  - bearerAuth: []
...
paths:
...
  /system/status:
    get:
      tags: [System]
      operationId: getSystemStatus
      summary: Degraded-mode system status
      responses:
        "200":
          description: Current operational status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /profile:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: Read profile
      description: >
        404 until onboarding writes one. Reading never creates a profile, so the
        client can tell "not onboarded yet" from "onboarded", and the device
        timezone it defaults to survives into the first PUT (R1).
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Profile]
      operationId: putProfile
      summary: Upsert profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileInput" }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /diet-profile:
    get:
      tags: [Profile]
      operationId: getDietProfile
      summary: Read diet profile
      responses:
        "200":
          description: Diet profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DietProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Profile]
      operationId: putDietProfile
      summary: Upsert diet profile (allergies are hard excludes)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DietProfileInput" }
      responses:
        "200":
          description: Updated diet profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DietProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /training-profile:
    get:
      tags: [Profile]
      operationId: getTrainingProfile
      summary: Read training profile
      responses:
        "200":
          description: Training profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Profile]
      operationId: putTrainingProfile
      summary: Upsert training profile (injuries are hard constraints)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrainingProfileInput" }
      responses:
        "200":
          description: Updated training profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /goals:
    get:
      tags: [Profile]
      operationId: listGoals
      summary: List goals (active first, ordered by priority)
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Goal" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [Profile]
      operationId: putGoal
      summary: Create or update a goal (upsert by id)
      description: >
        Upserts a single structured goal (R4). With an `id` it updates that goal;
        without one it creates a new goal. Multiple goals are supported by calling
        this repeatedly, each carrying its own `priority` (R14); a goal is retired
        by setting its `status` to paused/abandoned rather than a hard delete.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoalInput" }
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Goal" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }

  # ----------------------------------------------------------------- Settings
  /settings/api-keys:
    get:
      tags: [Profile]
      operationId: getApiKeyStatus
      summary: Provider API-key status (masked — never returns the secret)
      description: >
        Reports whether each provider key (Anthropic, Spoonacular) is set, with
        only a masked last-4 preview. The plaintext key is encrypted at rest and
        never returned after save (T1.3).
      responses:
        "200":
          description: Masked key status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [Profile]
      operationId: putApiKeys
      summary: Set / replace provider API keys
      description: >
        Stores the supplied provider key(s), encrypted at rest. Only provided,
        non-empty fields are updated. Returns the masked status — never the
        plaintext (T1.3).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiKeysInput" }
      responses:
        "200":
          description: Masked key status after the update
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }

  # ---------------------------------------------------------------- Training
  /training/program:generate:
    post:
      tags: [Training]
      summary: Generate a program (rules -> LLM -> validator)
      description: >
        Assembles training constraints from profile/goal/history, calls Claude
        with a schema-bound tool to select exercises within the allowed menu,
        validates volume landmarks and injury/equipment exclusions, and persists
        the Program with its input constraints.
      responses:
        "201":
          description: Program created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "422":
          description: Generation failed validation after repair attempts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerationError" }
  /training/program/current:
    get:
      tags: [Training]
      summary: Read the active program
      responses:
        "200":
          description: Current program
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "404": { $ref: "#/components/responses/NotFound" }
  /training/session/today:
    get:
      tags: [Training]
      summary: Today's session with pre-filled targets
      responses:
        "200":
          description: Today's session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkoutSession" }
        "404": { $ref: "#/components/responses/NotFound" }
  /training/session/{id}/log:
    post:
      tags: [Training]
      summary: Log actual sets for a session
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sets]
              properties:
                status:
                  type: string
                  enum: [completed, skipped, partial]
                sets:
                  type: array
                  items: { $ref: "#/components/schemas/SetLogInput" }
      responses:
        "200":
          description: Updated session with logged sets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkoutSession" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /training/session/{id}/feedback:
    post:
      tags: [Training]
      summary: Submit session feedback (free-text parsed into a signal)
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackInput" }
      responses:
        "202":
          description: Feedback accepted; will re-parameterize next generation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /training/progress:
    get:
      tags: [Training]
      summary: Progress series — volume, est-1RM, bodyweight
      parameters:
        - name: metric
          in: query
          schema: { type: string, enum: [volume, est1rm, bodyweight] }
        - name: exerciseId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Time series
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProgressSeries" }
  /exercises:
    get:
      tags: [Training]
      summary: Exercise library (filterable)
      parameters:
        - name: equipment
          in: query
          schema: { type: string }
        - name: muscle
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Exercise" }

  # ------------------------------------------------------------------- Meals
...
  /sync/push:
    post:
      tags: [Sync]
      summary: Push local changes (STUB — lands with its phase)
      description: >
        Idempotent by clientId. Body carries append-only events plus
        mutable-row upserts; the server appends one ChangeLog row per mutating
        write (R2) and replays are deduplicated by clientId (R2/R3). Syncable
        rows merge on the R3 sync quartet (version / deletedAt / clientId /
        updatedAt).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId:
                  type: string
                  description: Origin device id; makes the push idempotent (R2).
                events:
                  type: array
                  description: Append-only events (e.g. set logs, check-offs).
                  items: { type: object, additionalProperties: true }
                upserts:
                  type: array
                  description: Mutable-row upserts carrying the R3 sync quartet.
                  items: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Applied; returns the latest serverSeq cursor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  serverSeq:
                    type: integer
                    description: Highest ChangeLog serverSeq after applying this push.
                  applied: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/NotImplemented" }
  /sync/pull:
    get:
      tags: [Sync]
      summary: Pull changes since a cursor (STUB — lands with its phase)
      description: >
        Reads forward from the append-only ChangeLog (whose autoincrement PK is
        the monotonic serverSeq), returning changes and tombstones since the
        cursor (R2).
      parameters:
        - name: since
          in: query
          required: false
          schema: { type: integer, default: 0 }
          description: serverSeq cursor; returns rows with serverSeq > since (R2).
      responses:
        "200":
          description: Changes and tombstones since the cursor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  serverSeq:
                    type: integer
                    description: New cursor to persist for the next pull.
                  changes:
                    type: array
                    description: Upserted rows (table + row payload).
                    items: { type: object, additionalProperties: true }
                  tombstones:
                    type: array
                    description: Deleted rows ({ table, rowId }) since the cursor.
                    items: { type: object, additionalProperties: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/NotImplemented" }

  # ----------------------------------------------------------------- Devices
  /auth/tokens:
    get:
      tags: [Devices]
      operationId: listApiTokens
      summary: List per-device tokens (never returns the secret)
      description: >
        Lists minted device tokens with their metadata. The stored value is a
        SHA-256 hash; the plaintext secret is never returned after minting (T0.9).
      responses:
        "200":
          description: Device tokens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiToken" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Devices]
      operationId: mintApiToken
      summary: Mint a new per-device token
      description: >
        Mints a per-device bearer token. The response `token` is the plaintext
        secret and is shown ONLY once — only its SHA-256 hash is persisted (T0.9).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiTokenInput" }
      responses:
        "201":
          description: Token minted; `token` is the plaintext secret, shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MintedApiToken" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/tokens/{id}:
    delete:
      tags: [Devices]
      operationId: revokeApiToken
      summary: Revoke a per-device token
      description: >
        Revokes one device token by stamping `revokedAt`. From then on that token
        401s while every other token keeps working (T0.9). Idempotent.
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "204":
          description: Revoked (idempotent — already-revoked tokens also return 204).
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /pair:
    get:
      tags: [Devices]
      operationId: pairDevice
      summary: Redeem a pairing PIN for a per-device token (T0.12 · R22)
      description: >
        Device pairing. Requires an OPEN pairing window (opened by first-run
        `setup`) plus its valid short-lived PIN (R22). On success mints a new
        per-device bearer token and single-use-consumes the window. Outside any
        open window — or with a bad, expired, or already-consumed PIN — returns
        403; a trusted tailnet peer cannot silently pull a token (R22). This is
        the only unauthenticated endpoint (the device has no token yet); the PIN
        + open window are the gate.
      security: []
      parameters:
        - name: pin
          in: query
          required: true
          schema: { type: string }
          description: Short-lived pairing PIN carried by the setup QR (R22).
        - name: name
          in: query
          required: false
          schema: { type: string }
          description: >
            Optional device label for the minted token (e.g. "iPhone 15 Pro").
            Falls back to the window's configured device name.
      responses:
        "200":
          description: Paired; returns a per-device token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResult" }
        "403":
          description: No open pairing window, or invalid/expired/consumed PIN (R22).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Single static token over Tailscale.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request failed Zod validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    SystemStatus:
      type: object
      required: [mode, llm, provider, lastBackupAt, lastSyncAt, spendMTD, spendCeiling]
      properties:
        mode:
          type: string
          enum: [full, rules_local, rules_only]
          description: Degraded-mode ladder selected by current force toggles.
        llm:
          type: string
          enum: [up, down]
        provider:
          type: string
          enum: [up, down]
        lastBackupAt:
          type: string
          format: date-time
          nullable: true
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
        spendMTD:
          type: number
          description: Month-to-date LLM spend in USD.
        spendCeiling:
          type: number
          description: Configured monthly LLM spend ceiling in USD.
...
    Profile:
      type: object
      required:
        [id, age, sex, heightCm, weightKg, bodyFat, timezone, unitSystem, activityLevel]
      properties:
        id: { type: string }
        age: { type: integer, nullable: true }
        sex: { type: string, nullable: true }
        heightCm: { type: number, nullable: true }
        weightKg: { type: number, nullable: true }
        bodyFat: { type: number, nullable: true }
        timezone:
          type: string
          default: UTC
          description: IANA timezone (e.g. "America/New_York"); defines "today" (R1).
        unitSystem:
          type: string
          enum: [metric, imperial]
          default: metric
          description: Display preference only; storage is metric-canonical (R6).
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, very_active, athlete]
          default: moderate
          description: Seeds TDEE until the estimator supersedes it (R7).
    ProfileInput:
      type: object
      required: [timezone, unitSystem, activityLevel]
      properties:
        age: { type: integer }
        sex: { type: string, enum: [male, female, other] }
        heightCm: { type: number }
        weightKg: { type: number }
        bodyFat: { type: number }
        timezone:
          type: string
          default: UTC
          description: IANA timezone; defaulted from the device at onboarding (R1).
        unitSystem: { type: string, enum: [metric, imperial], default: metric }
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, very_active, athlete]
          default: moderate
...
    DietProfileInput:
      type: object
      properties:
        pattern: { type: string }
        restrictions: { type: array, items: { type: string } }
        allergies: { type: array, items: { type: string } }
        dislikes: { type: array, items: { type: string } }
        cuisines: { type: array, items: { type: string } }
        cookingSkill: { type: string, enum: [beginner, intermediate, advanced] }
        effortMax: { type: integer, minimum: 1, maximum: 5 }
        budgetWeekly:
          type: number
          description: Weekly food budget — a SOFT guide; the validator warns, not rejects (R12).
        mealsPerDay: { type: integer }
        snacksPerDay: { type: integer }
        batchCooking: { type: boolean }
        variety: { type: string, enum: [low, moderate, high] }
...
    ApiKeyStatus:
      type: object
      description: Masked status of every provider key. The secret is never returned (T1.3).
      required: [anthropic, spoonacular]
      properties:
        anthropic: { $ref: "#/components/schemas/ApiKeyState" }
        spoonacular: { $ref: "#/components/schemas/ApiKeyState" }
...